+++ example/recurring_example.dart
@@ -14,6 +14,7 @@ void main(List<String> args) async {
     task: () => getTime('Asia/Saigon'),
     interval: Duration(seconds: 3),
     maxRuns: 5,
+    queuePolicy: QueuePolicy.frontOfQueue,
   );
   DateTime cutoffTime = DateTime.now().add(Duration(days: 4));
   secretary.addRecurring(